consolidate FrequencySpectrum and SFS utilities into diversity.py
Move FrequencySpectrum class, compute_sigma_ij, project_sfs,
WEIGHT_REGISTRY, and all weight functions from achaz.py into
diversity.py. One module, one source of truth.

achaz.py becomes a thin backward-compatible re-export shim.
All 8 Achaz estimators validated at machine precision.
