+++ pg_gpu/diversity.py
@@ -3,13 +3,15 @@ GPU-accelerated diversity and polymorphism statistics.
 
 This module provides efficient computation of within-population genetic diversity
 metrics including nucleotide diversity (π), Watterson's theta, Tajima's D, and
-related statistics.
+related statistics. Includes the FrequencySpectrum class for SFS-based analysis,
+custom weight functions, and SFS projection.
 """
 
 import warnings
 import numpy as np
 import cupy as cp
-from typing import Union, Optional, Dict, Tuple
+from typing import Union, Optional, Dict, Tuple, Callable
+from functools import lru_cache
 from .haplotype_matrix import HaplotypeMatrix
 from ._utils import get_population_matrix
 
@@ -204,6 +206,400 @@ def _prepare_matrix(haplotype_matrix, population=None, missing_data='include'):
         matrix = matrix.exclude_missing_sites()
     return matrix
 
+
+# ---------------------------------------------------------------------------
+# SFS projection and covariance (Gutenkunst et al. 2009, Fu 1995)
+# ---------------------------------------------------------------------------
+
+@lru_cache(maxsize=64)
+def _harmonic_sums(n):
+    """Precompute harmonic sums H[i] = sum(1/j for j=1..i) for i=0..n."""
+    H = np.zeros(n + 1)
+    H[1:] = np.cumsum(1.0 / np.arange(1, n + 1))
+    return H
+
+
+@lru_cache(maxsize=64)
+def compute_sigma_ij(n):
+    """Compute the Fu (1995) covariance matrix sigma_ij for sample size n.
+
+    sigma_ij = Cov[xi_i, xi_j] / theta^2 for the unfolded SFS under the
+    standard neutral model.
+
+    Parameters
+    ----------
+    n : int
+        Sample size (number of haplotypes).
+
+    Returns
+    -------
+    sigma : ndarray, float64, shape (n-1, n-1)
+    """
+    H = _harmonic_sums(n)
+    sigma = np.zeros((n - 1, n - 1))
+
+    def beta(i, nn):
+        ai = H[i - 1]
+        an = H[nn - 1]
+        return 2.0 * nn / ((nn - i + 1) * (nn - i)) * (an + 1.0 / nn - ai) - 2.0 / (nn - i)
+
+    def sigma_ii(i, nn):
+        if 2 * i < nn:
+            return beta(i + 1, nn)
+        elif 2 * i == nn:
+            ai = H[i - 1]
+            an = H[nn - 1]
+            return 2.0 * (an - ai) / (nn - i) - 1.0 / (i * i)
+        else:
+            return beta(i, nn) - 1.0 / (i * i)
+
+    def sigma_ij_val(i, j, nn):
+        if i == j:
+            return sigma_ii(i, nn)
+        if i < j:
+            i, j = j, i
+        if i + j < nn:
+            return (beta(i + 1, nn) - beta(i, nn)) / 2.0
+        elif i + j == nn:
+            ai = H[i - 1]
+            aj = H[j - 1]
+            an = H[nn - 1]
+            return ((an - ai) / (nn - i) + (an - aj) / (nn - j)
+                    - (beta(i, nn) + beta(j + 1, nn)) / 2.0
+                    - 1.0 / (i * j))
+        else:
+            return (beta(j, nn) - beta(j + 1, nn)) / 2.0 - 1.0 / (i * j)
+
+    for i in range(1, n):
+        for j in range(1, n):
+            sigma[i - 1, j - 1] = sigma_ij_val(i, j, n)
+
+    return sigma
+
+
+@lru_cache(maxsize=128)
+def _projection_matrix(n_from, n_to):
+    """Hypergeometric projection matrix from n_from to n_to."""
+    from scipy.special import comb
+    P = np.zeros((n_to + 1, n_from + 1))
+    for k_from in range(n_from + 1):
+        for k_to in range(max(0, k_from - (n_from - n_to)),
+                          min(k_from, n_to) + 1):
+            P[k_to, k_from] = (comb(k_from, k_to, exact=True)
+                               * comb(n_from - k_from, n_to - k_to, exact=True)
+                               / comb(n_from, n_to, exact=True))
+    return P
+
+
+def project_sfs(sfs, n_from, n_to):
+    """Project an SFS from sample size n_from down to n_to.
+
+    Uses hypergeometric sampling (Gutenkunst et al. 2009).
+    """
+    if n_to > n_from:
+        raise ValueError(f"Cannot project up: n_to={n_to} > n_from={n_from}")
+    if n_to == n_from:
+        return sfs.copy()
+    return _projection_matrix(n_from, n_to) @ sfs
+
+
+# ---------------------------------------------------------------------------
+# FrequencySpectrum: power-user class for SFS analysis
+# ---------------------------------------------------------------------------
+
+# Weight functions for SFS dot-product path (used by FrequencySpectrum.theta
+# for custom callables; built-in names use _site_contribution instead).
+def _weights_watterson(n):
+    w = np.zeros(n + 1)
+    a1 = np.sum(1.0 / np.arange(1, n))
+    w[1:n] = 1.0 / a1
+    return w
+
+def _weights_pi(n):
+    k = np.arange(n + 1, dtype=np.float64)
+    w = 2.0 * k * (n - k) / (n * (n - 1))
+    w[0] = w[n] = 0.0
+    return w
+
+def _weights_theta_h(n):
+    k = np.arange(n + 1, dtype=np.float64)
+    w = 2.0 * k ** 2 / (n * (n - 1))
+    w[0] = w[n] = 0.0
+    return w
+
+def _weights_theta_l(n):
+    k = np.arange(n + 1, dtype=np.float64)
+    w = k / (n - 1)
+    w[0] = w[n] = 0.0
+    return w
+
+def _weights_eta1(n):
+    w = np.zeros(n + 1)
+    a1 = np.sum(1.0 / np.arange(1, n))
+    w[1] = 1.0 / a1
+    return w
+
+def _weights_eta1_star(n):
+    w = np.zeros(n + 1)
+    a1 = np.sum(1.0 / np.arange(1, n))
+    w[1] = w[n - 1] = 1.0 / a1
+    return w
+
+def _weights_minus_eta1(n):
+    w = np.zeros(n + 1)
+    a1 = np.sum(1.0 / np.arange(1, n))
+    w[2:n] = 1.0 / (a1 - 1.0)
+    return w
+
+def _weights_minus_eta1_star(n):
+    w = np.zeros(n + 1)
+    a1 = np.sum(1.0 / np.arange(1, n))
+    w[2:n - 1] = 1.0 / (a1 - 1.0 - 1.0 / (n - 1))
+    return w
+
+
+WEIGHT_REGISTRY: Dict[str, Callable] = {
+    'watterson': _weights_watterson, 'theta_s': _weights_watterson,
+    'pi': _weights_pi, 'theta_pi': _weights_pi,
+    'theta_h': _weights_theta_h,
+    'theta_l': _weights_theta_l,
+    'eta1': _weights_eta1,
+    'eta1_star': _weights_eta1_star,
+    'minus_eta1': _weights_minus_eta1,
+    'minus_eta1_star': _weights_minus_eta1_star,
+}
+
+
+class FrequencySpectrum:
+    """Site frequency spectrum with support for variable sample sizes.
+
+    Computes derived allele counts on GPU, groups by per-site sample
+    size, and provides theta estimation via weight vector dot products.
+    For built-in estimators, use the scalar functions (``pi()``, etc.)
+    which are faster. This class is for custom weight functions,
+    SFS inspection, projection, and the general Achaz variance framework.
+
+    Parameters
+    ----------
+    haplotype_matrix : HaplotypeMatrix
+    population : str or list, optional
+    missing_data : str
+        'include' (default) or 'exclude'
+    n_total_sites : int, optional
+        Total callable sites for invariant site correction.
+    """
+
+    def __init__(self, haplotype_matrix, population=None,
+                 missing_data='include', n_total_sites=None):
+        if population is not None:
+            matrix = get_population_matrix(haplotype_matrix, population)
+        else:
+            matrix = haplotype_matrix
+
+        if matrix.device == 'CPU':
+            matrix.transfer_to_gpu()
+        self._source_matrix = matrix
+        n_hap = matrix.num_haplotypes
+        if n_total_sites is None:
+            n_total_sites = matrix.n_total_sites
+
+        from ._memutil import dac_and_n as _dac_n
+        dac, n_valid = _dac_n(matrix.haplotypes)
+
+        if missing_data == 'exclude':
+            complete = n_valid == n_hap
+            dac = dac[complete]
+            n_valid = n_valid[complete]
+
+        self.sfs_by_n = {}
+        if len(dac) == 0:
+            self.n_max = 0
+            self.n_segregating = 0
+        else:
+            unique_n = cp.unique(n_valid)
+            self.n_max = int(unique_n[-1].get())
+            for ni_gpu in unique_n:
+                ni = int(ni_gpu.get())
+                if ni < 2:
+                    continue
+                mask = n_valid == ni
+                xi = cp.bincount(dac[mask], minlength=ni + 1)[:ni + 1]
+                self.sfs_by_n[ni] = xi.astype(cp.float64).get()
+            self.n_segregating = sum(
+                int(np.sum(xi[1:n])) for n, xi in self.sfs_by_n.items())
+
+        self.n_total_sites = n_total_sites
+        if n_total_sites is not None and self.n_max > 0:
+            n_invariant = n_total_sites - self.n_segregating
+            if n_invariant > 0 and self.n_max in self.sfs_by_n:
+                self.sfs_by_n[self.n_max][0] += n_invariant
+
+    def theta(self, weights='pi', span_normalize=False, span=None):
+        """Compute a theta estimator from the SFS.
+
+        Parameters
+        ----------
+        weights : str or callable
+            Name of a built-in weight function, or a callable w(n) -> array.
+        span_normalize : bool
+        span : float, optional
+        """
+        if isinstance(weights, str):
+            if weights not in WEIGHT_REGISTRY:
+                raise ValueError(f"Unknown weight: {weights}")
+            weights_fn = WEIGHT_REGISTRY[weights]
+        else:
+            weights_fn = weights
+
+        total = 0.0
+        for n, xi in self.sfs_by_n.items():
+            w = weights_fn(n)
+            total += np.sum(xi[:len(w)] * w[:len(xi)])
+
+        if span_normalize is not False:
+            if span is not None and span > 0:
+                total /= span
+            elif self._source_matrix is not None:
+                mode = 'auto' if span_normalize is True else span_normalize
+                s = self._source_matrix.get_span(mode)
+                if s > 0:
+                    total /= s
+
+        return total
+
+    def neutrality_test(self, w1='pi', w2='watterson'):
+        """Compute T = (theta1 - theta2) / sqrt(var) using Achaz (2009) Eq. 9."""
+        theta1 = self.theta(w1)
+        theta2 = self.theta(w2)
+        numerator = theta1 - theta2
+
+        n_eff = max(self.sfs_by_n.keys(),
+                    key=lambda n: np.sum(self.sfs_by_n[n]))
+
+        w1_fn = WEIGHT_REGISTRY[w1] if isinstance(w1, str) else w1
+        w2_fn = WEIGHT_REGISTRY[w2] if isinstance(w2, str) else w2
+        v1, v2 = w1_fn(n_eff), w2_fn(n_eff)
+
+        s1, s2 = np.sum(v1[1:n_eff]), np.sum(v2[1:n_eff])
+        if s1 == 0 or s2 == 0:
+            return float('nan')
+
+        Omega = np.array([v1[i]/s1 - v2[i]/s2 for i in range(1, n_eff)])
+        k = np.arange(1, n_eff, dtype=np.float64)
+        alpha_n = np.sum(k * Omega ** 2)
+
+        sigma = compute_sigma_ij(n_eff)
+        beta_n = float(np.sum(
+            k[:, None] * k[None, :] * Omega[:, None] * Omega[None, :] * sigma))
+
+        S = self.n_segregating
+        a1 = np.sum(1.0 / np.arange(1, n_eff))
+        a2 = np.sum(1.0 / np.arange(1, n_eff) ** 2)
+        theta_est = S / a1
+        theta_sq_est = S * (S - 1) / (a1 ** 2 + a2)
+
+        variance = alpha_n * theta_est + beta_n * theta_sq_est
+        if variance <= 0:
+            return float('nan')
+        return numerator / np.sqrt(variance)
+
+    def suggest_projection_n(self, retain_fraction=0.95):
+        """Suggest a projection target retaining most sites."""
+        if len(self.sfs_by_n) <= 1:
+            return self.n_max
+        sorted_ns = sorted(self.sfs_by_n.keys(), reverse=True)
+        total_seg = self.n_segregating
+        if total_seg == 0:
+            return self.n_max
+        cumulative = 0
+        for ni in sorted_ns:
+            cumulative += int(np.sum(self.sfs_by_n[ni][1:ni]))
+            if cumulative / total_seg >= retain_fraction:
+                return ni
+        return sorted_ns[-1]
+
+    def project(self, target_n):
+        """Project all SFS groups to a common sample size."""
+        projected = np.zeros(target_n + 1)
+        for n, xi in self.sfs_by_n.items():
+            if n < target_n:
+                continue
+            projected += project_sfs(xi, n, target_n)
+        result = object.__new__(FrequencySpectrum)
+        result.sfs_by_n = {target_n: projected}
+        result.n_max = target_n
+        result.n_segregating = int(np.sum(projected[1:target_n]))
+        result.n_total_sites = self.n_total_sites
+        result._source_matrix = self._source_matrix
+        return result
+
+    def sfs(self, n=None):
+        """Return the SFS, optionally projected."""
+        if n is not None:
+            return self.project(n).sfs_by_n[n]
+        if len(self.sfs_by_n) == 1:
+            return list(self.sfs_by_n.values())[0]
+        return self.sfs_by_n.get(self.n_max, np.array([]))
+
+    def all_thetas(self, span_normalize=False, span=None):
+        """Compute all 8 standard theta estimators."""
+        return {name: self.theta(name, span_normalize=span_normalize, span=span)
+                for name in ['pi', 'watterson', 'theta_h', 'theta_l',
+                             'eta1', 'eta1_star', 'minus_eta1', 'minus_eta1_star']}
+
+    def tajimas_d(self):
+        """Tajima's D using classical (1989) variance."""
+        pi = self.theta('pi')
+        S = float(self.n_segregating)
+        if S < 3:
+            return float('nan')
+        n = max(self.sfs_by_n.keys(),
+                key=lambda ni: np.sum(self.sfs_by_n[ni]))
+        e1, e2 = _tajimas_d_variance_coeffs(n)
+        a1 = np.sum(1.0 / np.arange(1, n))
+        tw = S / a1
+        d_var = e1 * S + e2 * S * (S - 1)
+        if d_var <= 0:
+            return float('nan')
+        return (pi - tw) / np.sqrt(d_var)
+
+    def fay_wu_h(self, normalized=False):
+        """Fay & Wu's H = pi - theta_H. Optionally normalized."""
+        h = self.theta('pi') - self.theta('theta_h')
+        if not normalized:
+            return h
+        n = max(self.sfs_by_n.keys(),
+                key=lambda ni: np.sum(self.sfs_by_n[ni]))
+        S = float(self.n_segregating)
+        a1 = np.sum(1.0 / np.arange(1, n))
+        a2 = np.sum(1.0 / np.arange(1, n) ** 2)
+        theta_est = S / a1
+        theta_sq = S * (S - 1) / (a1 ** 2 + a2)
+        tn = theta_est * (n - 2) / (6 * (n - 1))
+        tn2 = theta_sq * (18 * n**2 * (3*n+2) * a2
+                          - (88*n**3 + 9*n**2 - 13*n + 6)) / (9*n*(n-1)**2)
+        var_h = tn + tn2
+        if var_h <= 0:
+            return float('nan')
+        return h / np.sqrt(var_h)
+
+    def zeng_e(self):
+        """Zeng's E via Achaz general variance framework."""
+        return self.neutrality_test('theta_l', 'watterson')
+
+    def all_tests(self):
+        """All standard neutrality tests."""
+        return {
+            'tajimas_d': self.tajimas_d(),
+            'fay_wu_h': self.fay_wu_h(),
+            'normalized_fay_wu_h': self.fay_wu_h(normalized=True),
+            'zeng_e': self.zeng_e(),
+        }
+
+
+# ---------------------------------------------------------------------------
+# Pairwise components (power-user API)
+# ---------------------------------------------------------------------------
 
 def pi_components(haplotypes, n_total_sites=None, n_haplotypes_full=None):
     """Compute pairwise differences and comparisons across all sites.
